third_party/WebKit/LayoutTests/fast/dom/timer-clear-interval-in-handler-and-generate-error-expected.txt

CONSOLE MESSAGE: line 10: Uncaught NotFoundError: An attempt was made to reference a Node in a context where it does not exist.
